passes: Refactor Parallelizer hierarchy
