+++ devito/passes/iet/__init__.py
@@ -1,7 +1,7 @@
 from .engine import *  # noqa
-from .openmp import *  # noqa
 from .orchestration import *  # noqa
 from .mpi import *  # noqa
 from .misc import *  # noqa
 from .definitions import *  # noqa
 from .instrument import *  # noqa
+from .specializations import *  # noqa
