+++ devito/passes/iet/specializations/C.py
@@ -1,0 +1,12 @@
+from devito.passes.iet.language import LanguageSpecializer
+
+__all__ = ['Cizer']
+
+
+class Cizer(LanguageSpecializer):
+
+    """
+    A dummy specializer with no-op transformers.
+    """
+
+    pass
