+++ devito/passes/iet/specializations/__init__.py
@@ -1,0 +1,2 @@
+from .C import *  # noqa
+from .openmp import *  # noqa
