+++ tests/test_dle.py
@@ -8,7 +8,7 @@ from devito import (Grid, Function, TimeFunction, SparseTimeFunction, SpaceDimen
                     Dimension, SubDimension, Eq, Inc, Operator, info)
 from devito.exceptions import InvalidArgument
 from devito.ir.iet import Call, Iteration, Conditional, FindNodes, retrieve_iteration_tree
-from devito.passes.iet.openmp import OpenMPRegion
+from devito.passes.iet import OpenMPRegion
 from devito.tools import as_tuple
 from devito.types import Scalar, NThreads, NThreadsNonaffine
